+++ packages/eslint-config-peppy/prettier.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ["./rules/prettier.js"].map(require.resolve),
+};
