--- packages/eslint-config-peppy/rules/react-ts.js
@@ -1,13 +1,0 @@
-module.exports = {
-  settings: {
-    "import/resolver": {
-      node: {
-        extensions: [".mjs", ".js", ".jsx", ".json", ".ts", ".tsx", ".d.ts"],
-      },
-    },
-  },
-  rules: {
-    "react/jsx-filename-extension": ["error", { extensions: [".jsx", ".tsx"] }],
-    "react/prop-types": "off",
-  },
-};
