--- packages/eslint-config-peppy/ts.js
@@ -1,5 +1,0 @@
-module.exports = {
-  extends: ["./rules/base.js", "./rules/base-ts.js", "./rules/prettier.js"].map(
-    require.resolve
-  ),
-};
